create a submodule for arbor's dsl inside parameterscalers
